requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
customtkinter==5.1.2
darkdetect==0.8.0
decorator==5.1.1
future==0.18.3
geocoder==1.38.1
idna==3.4
Pillow==9.5.0
pyperclip==1.8.2
pywin32==306
ratelim==0.1.6
requests==2.28.2
six==1.16.0
tkintermapview==1.29
urllib3==1.26.15
